requirements.txt
# requirements-minimal.txt
# Minimal dependencies for retail shelf captioning

# Core PyTorch (CPU only)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0

# Data science basics
numpy>=1.24.0
pandas>=2.0.0

# Evaluation metrics
nltk>=3.8.0
rouge-score>=0.1.2

# Training utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# CLI and file handling
click>=8.1.0
requests>=2.28.0
gdown>=4.6.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
